requirements.txt
python-jose[cryptography]==3.3.0
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
passlib==1.7.4
python-multipart ==25.0.1
bcrypt ==25.0.1